requirements.txt
feedparser
requests
Pillow
atproto